backend/requirement.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
Django==3.2.1
django-cors-headers==3.11.0
django-rest-knox==4.1.0
djangorestframework==3.13.1
Pillow==9.2.0
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
tzdata==2022.1
